fix: u32 overflow in stitch offsets + avoid clone under mutex
Cast to usize before multiplying row offsets to prevent u32 overflow
on long scroll captures with 5K displays (y * width * 4 can exceed
4.2B). Use std::mem::take instead of clone() when draining frames
for stitching to halve peak memory usage.

+++ app/src-tauri/src/stitch.rs
@@ -113,10 +113,10 @@ pub fn stitch_frames(frames: Vec<DynamicImage>) -> Result<String, String> {
         // Row-wise copy: RgbaImage stores rows contiguously (RGBA = 4 bytes/pixel).
         // copy_from_slice is ~10x faster than per-pixel put_pixel (eliminates
         // bounds checks and function call overhead per pixel).
-        let bytes_per_row = (frame_w * 4) as usize;
+        let bytes_per_row = (frame_w as usize) * 4;
         for y in skip..src.height() {
-            let src_offset = (y * src.width() * 4) as usize;
-            let dst_offset = (y_out * width * 4) as usize;
+            let src_offset = (y as usize) * (src.width() as usize) * 4;
+            let dst_offset = (y_out as usize) * (width as usize) * 4;
             result.as_mut()[dst_offset..dst_offset + bytes_per_row]
                 .copy_from_slice(&src.as_raw()[src_offset..src_offset + bytes_per_row]);
             y_out += 1;
